Don't crash update when we're already up to date
Trimming the lifetime intervals fails when the latest price is newer
that the end date of the update. Avoid trimming the intervals in this
case and avoid updating altogether as well.

This can happen e.g. when we've already updated the prices today, so the
date of the latest price and the end date of the update are both equal
to the current date.