Fix cache clearing and improve documentation.
Cache clearing did not work properly, because `shelve` library can add
additional extensions to the filename to store the cache database (e.g.
it adds `.db` extension in my case). This means that removing the
`cache_filename` file does not work, as such file is not actually
created.

Instead of removing the files manually, I am using flag='n' which
ensures that the cache will be cleared.

I also removed the unit tests that depended on `setup_cache` to use
`os.remove`, and added a new E2E test that ensures that when we clear
the cache, the old values are actually not used (the new test fails on
old implementation of `setup_cache`).

I also updated documentation (comments and flag descriptions).